Move Linux code to new respository structure
