+++ example/linux/.gitignore
