--- linux/Makefile
@@ -1,28 +1,0 @@
-# Copyright 2018 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-TARGETS=all clean
-PLUGINS=color_panel file_chooser menubar
-PLUGIN_DIRS=$(patsubst %,../plugins/%/linux,$(PLUGINS))
-COMPONENT_DIRS=library example $(PLUGIN_DIRS)
-
-# Enforce relative ordering. This prevents accidentally building the targets
-# in parallel in the event that someone supplies the -j flag to `make`.
-example: library $(PLUGIN_DIRS)
-$(PLUGIN_DIRS): library
-
-$(TARGETS): $(COMPONENT_DIRS)
-$(COMPONENT_DIRS):
-	$(MAKE) -C $@ $(MAKECMDGOALS)
-
-.PHONY: $(TARGETS) $(COMPONENT_DIRS)
